delete view for the list

+++ Todo/MyToDo/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render,redirect,HttpResponse
+from django.shortcuts import render,redirect,HttpResponse,get_object_or_404
 from django.contrib import messages
 
 # import todo form and models
@@ -10,7 +10,7 @@ from .models import Todo
 # Create your views here.
 
 
-def index(request):
+def add_item(request):
     item_list = Todo.objects.order_by("-date")
     if request.method == "POST":
         form = TodoForm(request.POST)
@@ -19,7 +19,7 @@ def index(request):
             return HttpResponse('Details submitted successfully')
     form = TodoForm()
     page = {
-        "forms":form,
+        "todo_form":form,
         "list":item_list,
         "title":"TODO LIST"
     }
@@ -27,14 +27,20 @@ def index(request):
     return render(request,"MyToDo/index.html",page)
 
 
-
+  ####################give id no. item_id name or item_id=i.id ############
 ### function to remove item, it receive todo item_id as primary key from url ##
 
-def remove(request,item_id):
-    item = Todo.objects.get(id=item_id)
-    item.delete()
-    messages.info(request,"Item removed !!!")
-    return redirect("todo")
+def remove_item(request,item_id):
+    print("Received request to remove item with ID:", item_id)
+    print("Current Todo IDs:", Todo.objects.values_list('id', flat=True))  # Log current IDs
+    if request.method == "POST":
+        item = get_object_or_404(Todo,id=item_id)  # Automatically handles DoesNotExist
+        item.delete()
+        messages.info(request, "Item removed!")
+    else:
+        messages.error(request, "Invalid request method.")
+    return redirect("todo:add")
+
